PCG_Predict_Models/requirements.txt
torch>=2.1.0
torch-geometric>=2.4.0
onnx>=1.15.0
onnxruntime>=1.16.0
pyyaml>=6.0
numpy>=1.24.0
pytest>=7.4.0
